Rename: src/lib/install-dependencies.js -> src/lib/setup.js
